Hide dedicated sessions behind feature toggle
