+++ graphdatascience/session/__init__.py
@@ -1,7 +1,8 @@
 from .algorithm_category import AlgorithmCategory
 from .dbms_connection_info import DbmsConnectionInfo
-from .gds_sessions import AuraAPICredentials, GdsSessions, SessionInfo
+from .gds_sessions import AuraAPICredentials, GdsSessions
 from .schema import GdsPropertyTypes
+from .session_info import SessionInfo
 from .session_sizes import SessionMemory
 
 __all__ = [
