+++ graphdatascience/tests/unit/test_aurads_sessions.py
@@ -1,11 +1,9 @@
 import dataclasses
 import re
-from dataclasses import asdict
 from typing import Any, List, Optional
 
 import pytest
 from pytest_mock import MockerFixture
-from requests_mock import Mocker
 
 from graphdatascience.session.algorithm_category import AlgorithmCategory
 from graphdatascience.session.aura_api import AuraApi
@@ -17,13 +15,9 @@ from graphdatascience.session.aura_api_responses import (
     TenantDetails,
     WaitResult,
 )
+from graphdatascience.session.aurads_sessions import AuraDsSessions, SessionNameHelper
 from graphdatascience.session.dbms_connection_info import DbmsConnectionInfo
-from graphdatascience.session.gds_sessions import (
-    AuraAPICredentials,
-    GdsSessionNameHelper,
-    GdsSessions,
-    SessionInfo,
-)
+from graphdatascience.session.session_info import SessionInfo
 from graphdatascience.session.session_sizes import SessionMemory
 
 
@@ -103,33 +97,11 @@ def aura_api() -> AuraApi:
     return FakeAuraApi()
 
 
-def test_list_session(requests_mock: Mocker) -> None:
-    sessions = GdsSessions(AuraAPICredentials("", "", "placeholder"))
+def test_list_session(aura_api: AuraApi) -> None:
+    aura_api = FakeAuraApi()
+    aura_api.create_instance(name="gds-session-my-session-name", cloud_provider="gcp", memory="16GB", region="")
 
-    db_instance = InstanceDetails(id="id", name="Instance01", tenant_id="tenant_id", cloud_provider="cloud_provider")
-    gds_instance = InstanceDetails(
-        id="id", name="gds-session-my-session-name", tenant_id="tenant_id", cloud_provider="cloud_provider"
-    )
-    gds_instance_specific_details = InstanceSpecificDetails(
-        id="id",
-        name="gds-session-my-session-name",
-        tenant_id="tenant_id",
-        cloud_provider="gcp",
-        status="RUNNING",
-        connection_url="",
-        memory="16GB",
-        type="gds",
-        region="",
-    )
-
-    requests_mock.post(
-        "https://api.neo4j.io/oauth/token",
-        json={"access_token": "my_token", "expires_in": 3600, "token_type": "Bearer"},
-    )
-    requests_mock.get(
-        "https://api.neo4j.io/v1/instances", json={"data": [asdict(i) for i in [db_instance, gds_instance]]}
-    )
-    requests_mock.get("https://api.neo4j.io/v1/instances/id", json={"data": asdict(gds_instance_specific_details)})
+    sessions = AuraDsSessions(aura_api)
 
     assert sessions.list() == [SessionInfo("my-session-name", "16GB")]
 
@@ -137,14 +109,15 @@ def test_list_session(requests_mock: Mocker) -> None:
 def test_create_session(mocker: MockerFixture, aura_api: AuraApi) -> None:
     _setup_db_instance(aura_api)
 
-    sessions = GdsSessions(AuraAPICredentials("", "", "placeholder"))
-    sessions._aura_api = aura_api
+    sessions = AuraDsSessions(aura_api)
 
     def assert_db_credentials(*args: List[Any], **kwargs: str) -> None:
         assert kwargs == {"gds_url": "fake-url", "gds_user": "neo4j", "initial_pw": "fake-pw", "new_pw": "db_pw"}
 
-    mocker.patch("graphdatascience.session.gds_sessions.GdsSessions._construct_client", lambda *args, **kwargs: kwargs)
-    mocker.patch("graphdatascience.session.gds_sessions.GdsSessions._change_initial_pw", assert_db_credentials)
+    mocker.patch(
+        "graphdatascience.session.aurads_sessions.AuraDsSessions._construct_client", lambda *args, **kwargs: kwargs
+    )
+    mocker.patch("graphdatascience.session.aurads_sessions.AuraDsSessions._change_initial_pw", assert_db_credentials)
 
     gds_credentials = sessions.get_or_create(
         "my-session",
@@ -170,11 +143,14 @@ def test_create_session(mocker: MockerFixture, aura_api: AuraApi) -> None:
 def test_create_default_session(mocker: MockerFixture, aura_api: AuraApi) -> None:
     _setup_db_instance(aura_api)
 
-    sessions = GdsSessions(AuraAPICredentials("", "", "placeholder"))
-    sessions._aura_api = aura_api
+    sessions = AuraDsSessions(aura_api)
 
-    mocker.patch("graphdatascience.session.gds_sessions.GdsSessions._construct_client", lambda *args, **kwargs: kwargs)
-    mocker.patch("graphdatascience.session.gds_sessions.GdsSessions._change_initial_pw", lambda *args, **kwargs: kwargs)
+    mocker.patch(
+        "graphdatascience.session.aurads_sessions.AuraDsSessions._construct_client", lambda *args, **kwargs: kwargs
+    )
+    mocker.patch(
+        "graphdatascience.session.aurads_sessions.AuraDsSessions._change_initial_pw", lambda *args, **kwargs: kwargs
+    )
 
     sessions.get_or_create(
         "my-session",
@@ -190,11 +166,14 @@ def test_create_default_session(mocker: MockerFixture, aura_api: AuraApi) -> Non
 def test_create_session_override_region(mocker: MockerFixture, aura_api: AuraApi) -> None:
     aura_api.create_instance("test", "8GB", "aws", "dresden-2")
 
-    sessions = GdsSessions(AuraAPICredentials("", "", "placeholder"))
-    sessions._aura_api = aura_api
+    sessions = AuraDsSessions(aura_api)
 
-    mocker.patch("graphdatascience.session.gds_sessions.GdsSessions._construct_client", lambda *args, **kwargs: kwargs)
-    mocker.patch("graphdatascience.session.gds_sessions.GdsSessions._change_initial_pw", lambda *args, **kwargs: kwargs)
+    mocker.patch(
+        "graphdatascience.session.aurads_sessions.AuraDsSessions._construct_client", lambda *args, **kwargs: kwargs
+    )
+    mocker.patch(
+        "graphdatascience.session.aurads_sessions.AuraDsSessions._change_initial_pw", lambda *args, **kwargs: kwargs
+    )
 
     sessions.get_or_create(
         "my-session",
@@ -210,11 +189,14 @@ def test_create_session_override_region(mocker: MockerFixture, aura_api: AuraApi
 def test_get_or_create(mocker: MockerFixture, aura_api: AuraApi) -> None:
     _setup_db_instance(aura_api)
 
-    sessions = GdsSessions(AuraAPICredentials("", "", "placeholder"))
-    sessions._aura_api = aura_api
+    sessions = AuraDsSessions(aura_api)
 
-    mocker.patch("graphdatascience.session.gds_sessions.GdsSessions._construct_client", lambda *args, **kwargs: kwargs)
-    mocker.patch("graphdatascience.session.gds_sessions.GdsSessions._change_initial_pw", lambda *args, **kwargs: None)
+    mocker.patch(
+        "graphdatascience.session.aurads_sessions.AuraDsSessions._construct_client", lambda *args, **kwargs: kwargs
+    )
+    mocker.patch(
+        "graphdatascience.session.aurads_sessions.AuraDsSessions._change_initial_pw", lambda *args, **kwargs: None
+    )
 
     gds_args1 = sessions.get_or_create(
         "my-session",
@@ -242,11 +224,14 @@ def test_get_or_create(mocker: MockerFixture, aura_api: AuraApi) -> None:
 def test_get_or_create_different_size(mocker: MockerFixture, aura_api: AuraApi) -> None:
     _setup_db_instance(aura_api)
 
-    sessions = GdsSessions(AuraAPICredentials("", "", "placeholder"))
-    sessions._aura_api = aura_api
+    sessions = AuraDsSessions(aura_api)
 
-    mocker.patch("graphdatascience.session.gds_sessions.GdsSessions._construct_client", lambda *args, **kwargs: kwargs)
-    mocker.patch("graphdatascience.session.gds_sessions.GdsSessions._change_initial_pw", lambda *args, **kwargs: None)
+    mocker.patch(
+        "graphdatascience.session.aurads_sessions.AuraDsSessions._construct_client", lambda *args, **kwargs: kwargs
+    )
+    mocker.patch(
+        "graphdatascience.session.aurads_sessions.AuraDsSessions._change_initial_pw", lambda *args, **kwargs: None
+    )
 
     sessions.get_or_create(
         "my-session",
@@ -268,12 +253,11 @@ def test_get_or_create_different_size(mocker: MockerFixture, aura_api: AuraApi) 
 
 
 def test_get_or_create_duplicate_session() -> None:
-    sessions = GdsSessions(AuraAPICredentials("", "", "placeholder"))
-    sessions._aura_api = FakeAuraApi(
+    aura_api = FakeAuraApi(
         existing_instances=[
             InstanceSpecificDetails(
                 id="id42",
-                name=GdsSessionNameHelper.instance_name("one"),
+                name=SessionNameHelper.instance_name("one"),
                 tenant_id="",
                 cloud_provider="",
                 status="RUNNING",
@@ -284,7 +268,7 @@ def test_get_or_create_duplicate_session() -> None:
             ),
             InstanceSpecificDetails(
                 id="id43",
-                name=GdsSessionNameHelper.instance_name("one"),
+                name=SessionNameHelper.instance_name("one"),
                 tenant_id="",
                 cloud_provider="",
                 status="RUNNING",
@@ -296,17 +280,17 @@ def test_get_or_create_duplicate_session() -> None:
         ]
     )
 
+    sessions = AuraDsSessions(aura_api)
+
     with pytest.raises(RuntimeError, match=re.escape("Expected to find exactly one GDS session with name `one`")):
         sessions.get_or_create("one", SessionMemory.m_8GB, DbmsConnectionInfo("", "", ""))
 
 
 def test_delete_session() -> None:
-    sessions = GdsSessions(AuraAPICredentials("", "", "placeholder"))
-
     existing_instances = [
         InstanceSpecificDetails(
             id="id42",
-            name=GdsSessionNameHelper.instance_name("one"),
+            name=SessionNameHelper.instance_name("one"),
             tenant_id="tenant_id",
             cloud_provider="cloud_provider",
             status="RUNNING",
@@ -317,7 +301,7 @@ def test_delete_session() -> None:
         ),
         InstanceSpecificDetails(
             id="id1",
-            name=GdsSessionNameHelper.instance_name("other"),
+            name=SessionNameHelper.instance_name("other"),
             tenant_id="tenant_id",
             cloud_provider="cloud_provider",
             status="RUNNING",
@@ -328,19 +312,17 @@ def test_delete_session() -> None:
         ),
     ]
 
-    sessions._aura_api = FakeAuraApi(existing_instances=existing_instances)
+    sessions = AuraDsSessions(FakeAuraApi(existing_instances=existing_instances))
 
     assert sessions.delete("one")
     assert sessions.list() == [SessionInfo("other", "")]
 
 
 def test_delete_nonexisting_session() -> None:
-    sessions = GdsSessions(AuraAPICredentials("", "", "placeholder"))
-
     existing_instances = [
         InstanceSpecificDetails(
             id="id42",
-            name=GdsSessionNameHelper.instance_name("one"),
+            name=SessionNameHelper.instance_name("one"),
             tenant_id="tenant_id",
             cloud_provider="cloud_provider",
             status="RUNNING",
@@ -351,19 +333,17 @@ def test_delete_nonexisting_session() -> None:
         ),
     ]
 
-    sessions._aura_api = FakeAuraApi(existing_instances=existing_instances)
+    sessions = AuraDsSessions(FakeAuraApi(existing_instances=existing_instances))
 
     assert sessions.delete("other") is False
     assert sessions.list() == [SessionInfo("one", "")]
 
 
 def test_delete_nonunique_session() -> None:
-    sessions = GdsSessions(AuraAPICredentials("", "", "placeholder"))
-
     existing_instances = [
         InstanceSpecificDetails(
             id="id42",
-            name=GdsSessionNameHelper.instance_name("one"),
+            name=SessionNameHelper.instance_name("one"),
             tenant_id="",
             cloud_provider="",
             status="RUNNING",
@@ -374,7 +354,7 @@ def test_delete_nonunique_session() -> None:
         ),
         InstanceSpecificDetails(
             id="id43",
-            name=GdsSessionNameHelper.instance_name("one"),
+            name=SessionNameHelper.instance_name("one"),
             tenant_id="",
             cloud_provider="",
             status="RUNNING",
@@ -385,8 +365,7 @@ def test_delete_nonunique_session() -> None:
         ),
     ]
 
-    sessions._aura_api = FakeAuraApi(existing_instances=existing_instances)
-
+    sessions = AuraDsSessions(FakeAuraApi(existing_instances=existing_instances))
     with pytest.raises(
         RuntimeError,
         match=re.escape(
@@ -402,8 +381,7 @@ def test_delete_nonunique_session() -> None:
 def test_create_immediate_delete(aura_api: AuraApi) -> None:
     aura_api = FakeAuraApi(status_after_creating="deleting")
     _setup_db_instance(aura_api)
-    sessions = GdsSessions(AuraAPICredentials("", "", "foo"))
-    sessions._aura_api = aura_api
+    sessions = AuraDsSessions(aura_api)
 
     with pytest.raises(RuntimeError, match="Failed to create session `one`: Instance is being deleted"):
         sessions.get_or_create(
@@ -414,8 +392,7 @@ def test_create_immediate_delete(aura_api: AuraApi) -> None:
 def test_create_waiting_forever() -> None:
     aura_api = FakeAuraApi(status_after_creating="updating")
     _setup_db_instance(aura_api)
-    sessions = GdsSessions(AuraAPICredentials("", "", "foo"))
-    sessions._aura_api = aura_api
+    sessions = AuraDsSessions(aura_api)
 
     with pytest.raises(RuntimeError, match="Failed to create session `one`: Instance is not running after waiting"):
         sessions.get_or_create(
@@ -426,8 +403,7 @@ def test_create_waiting_forever() -> None:
 def test_invalid_session_name() -> None:
     aura_api = FakeAuraApi(status_after_creating="updating")
     _setup_db_instance(aura_api)
-    sessions = GdsSessions(AuraAPICredentials("", "", "foo"))
-    sessions._aura_api = aura_api
+    sessions = AuraDsSessions(aura_api)
 
     with pytest.raises(ValueError, match="Session name `one_very_long_session_name` is too long. Max length is 18."):
         sessions.get_or_create(
@@ -437,48 +413,22 @@ def test_invalid_session_name() -> None:
         )
 
 
-def test_estimate_size(aura_api: AuraApi) -> None:
+def test_estimate_size() -> None:
     aura_api = FakeAuraApi(size_estimation=EstimationDetails("1GB", "8GB", False))
-    sessions = GdsSessions(AuraAPICredentials("", "", "foo"))
-    sessions._aura_api = aura_api
+    sessions = AuraDsSessions(aura_api)
 
     assert sessions.estimate(1, 1, [AlgorithmCategory.CENTRALITY]) == SessionMemory.m_8GB
 
 
-def test_estimate_size_exceeds(aura_api: AuraApi) -> None:
+def test_estimate_size_exceeds() -> None:
     aura_api = FakeAuraApi(size_estimation=EstimationDetails("16GB", "8GB", True))
-    sessions = GdsSessions(AuraAPICredentials("", "", "foo"))
-    sessions._aura_api = aura_api
+    sessions = AuraDsSessions(aura_api)
 
     with pytest.warns(
         ResourceWarning,
         match=re.escape("The estimated memory `16GB` exceeds the maximum size supported by your Aura tenant (`8GB`)"),
     ):
         assert sessions.estimate(1, 1, [AlgorithmCategory.CENTRALITY]) == SessionMemory.m_8GB
-
-
-def test_from_specific_instance_details() -> None:
-    info = SessionInfo.from_specific_instance_details(
-        InstanceSpecificDetails("id", "gds-session-foo", "tenant", "cp", "status", "bolt", "96GB", "type", "region")
-    )
-
-    assert info.name == "foo"
-    assert info.memory == "96GB"
-
-
-def test_from_specific_instance_details_failures() -> None:
-    with pytest.raises(ValueError, match="Unknown memory configuration: `invalid`. Supported values are `\\['8GB', "):
-        SessionInfo.from_specific_instance_details(
-            InstanceSpecificDetails(
-                "id", "gds-session-foo", "tenant", "cp", "status", "bolt", "invalid", "type", "region"
-            )
-        )
-    with pytest.raises(ValueError, match="Invalid session name: `not-a-gds-session-foo`"):
-        SessionInfo.from_specific_instance_details(
-            InstanceSpecificDetails(
-                "id", "not-a-gds-session-foo", "tenant", "cp", "status", "bolt", "24GB", "type", "region"
-            )
-        )
 
 
 def _setup_db_instance(aura_api: AuraApi) -> None:
